% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDists.R
\name{compDists}
\alias{compDists}
\title{Compute distances between Peaks}
\usage{
compDists(MD, dist.method = "MMD", sigma = NULL, CompIDs = NULL,
  strand.specific = FALSE, run.parallel = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{dist.method}{specify method used for distances
between samples. Currently only Maximum Mean Discrepancy (MMD)
and Kolmogorov-Smirnov (KS) implemented.
(DEFAULT: 'MMD')}

\item{sigma}{sigma parameter of the RBF kernel,
determining the distance (along the genome) at which fragment counts
decorrelate. If set to NULL, 100 random Peaks are used to determine sigma
heuristically according to the method described in the MMDiff paper [1].
(DEFUALT: NULL)}

\item{CompIDs}{(2 x C) Matrix specifying for pairs of Sample IDs for which
distances should be computed. If NULL all pairwise distances will be computed.
(DEFAULT: NULL)}

\item{strand.specific}{if strand specific reads should be kept apart.
(DEFAULT:FALSE)}

\item{run.parallel}{whether to run in parallel
(currently no parallelization implemented)
(DEFAULT: FALSE)}

\item{BPPARAM}{an optional parameter object passed internally
to \code{\link{bplapply}} when \code{parallel=TRUE}.
If not specified, the parameters last registered with
\code{\link{register}} will be used.}
}
\value{
DBAmmd object with updated slot \code{Dists}
}
\description{
This function computes pairwise distances between histograms
according to the dist.method (MMD, KS). For large data sets it is a bit
time consuming.
}
\examples{

## Example using a small data set provided with this package:

data("MMD")
MMD.1 <- compDists(MMD)

# To inspect the computed distances:
D <- Dists(MMD.1,dist.method='MMD')
head(D)

# To analyse the result:
plotDists(MMD.1)

}
\references{
[1] Schweikert et al.  BMC Genomics 2013
...
}
\seealso{
\code{\link{DBAmmd}}, \code{\link{plotDists}},
\code{\link{plotDISTS4Peak}}, \code{\link{compPvals}}
}
\author{
Gabriele Schweikert \email{G.Schweikert@ed.ac.uk}
}
