% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTSContrast.R
\name{createTSContrast}
\alias{createTSContrast}
\title{createTSContrast}
\usage{
createTSContrast(MD, condition = NULL, tp = NULL)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{condition}{(DEFAULT: NULL)}

\item{tp}{time points (DEFAULT: NULL)}
}
\description{
This function creates all contrasts for a time series experiments,
comparing all time points tp[i] with following time point tp[i+1].
If condition is set, only samples with given condition are considered.
}
