% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTSContrast.R
\name{createTSContrastBycondition}
\alias{createTSContrastBycondition}
\title{createTSContrastBycondition}
\usage{
createTSContrastBycondition(MD, conditions = NULL, tp = NULL)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{conditions}{If NULL first two conditions are considered. (DEFAULT: NULL)}

\item{tp}{time points. If NULL all available time points are considred
(DEFAULT: NULL)}
}
\description{
This function creates contrasts between conditions[1] and condistions[2] for
all time points tp of a time series experiments.
}
