% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMDiff2_data.R
\docType{data}
\name{mm9-Genes}
\alias{mm9-Genes}
\title{mm9-Genes}
\format{contains, GR a GRanges object with 800 ranges}
\usage{
data('mm9-Genes')
}
\description{
Subset of Genes from the mm9 annotation that overlap with example Peaks in
the Cfp1-Peaks file.
}
\examples{
# data was created as follows:
\dontrun{
data('Cfp1-Peaks')
library(TxDb.Mmusculus.UCSC.mm9.knownGene)
txdb <- TxDb.Mmusculus.UCSC.mm9.knownGene #shorthand (for convenience) txdb
GR <- transcripts(txdb)
ov <- findOverlaps(GR,Peaks)
GR <- GR[queryHits(ov)]
save(file = 'data/mm9-Genes.rData',GR)
}


}
\keyword{data}
