% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportResults.R
\name{reportResults}
\alias{reportResults}
\title{report results}
\usage{
reportResults(MD, diff.method = "MMD.locfit", th = 0.1, Peak.id = NULL,
  whichContrast = 1, rm.oulier = TRUE, bUsePval = FALSE)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{diff.method}{method used to determine p-values and false discovery rates.
Currently only 'MMD.locfit' implemented.
(DEFAULT: 'MMD.locfit')}

\item{th}{significance threshold for differential called peaks (DEFAULT: 0.1)}

\item{Peak.id}{Peak id to specify which Peak to plot.
(coresponding to names of Regions(MD))}

\item{whichContrast}{index determining which of the set contrast should be used.
(DEFAULT: 1)}

\item{rm.oulier}{if TRUE, significant peaks with high
within-group distances are not reported. (DEFAULT: TRUE)}

\item{bUsePval}{if TRUE p-values instead of FDRs are used (DEFAULT: FALSE)}
}
\description{
retrieve results of differential binding analysis
}
\examples{

data("MMD")
res <- reportResults(MMD)

}
