% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server.MMDiff2}
\alias{server.MMDiff2}
\title{Shiny server code for interactive visualization of MMD distances,
peak plots, and MMD distances by sample.}
\usage{
server.MMDiff2(MD, whichContrast = 1)
}
\arguments{
\item{MD}{DBAmmd Object. This Object can be created using \code{DBAmmd()}.}

\item{whichContrast}{index determining which of the set contrast should be used.
(DEFAULT: 1)}
}
\description{
Shiny server code for interactive visualization of MMD distances,
peak plots, and MMD distances by sample.
}
\examples{
 if(interactive()){
    data("MMD")
    runShinyMMDiff2(MMD)
}

}
