% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui.MMDiff2}
\alias{ui.MMDiff2}
\title{ui component for interactive visualization of MMD,GMD and Pearson Difference
as well as plotting peaks}
\usage{
ui.MMDiff2(MD)
}
\arguments{
\item{MD}{DBAmmd object}
}
\description{
ui component for interactive visualization of MMD,GMD and Pearson Difference
as well as plotting peaks
}
\examples{
 if(interactive()){
load(system.file("data/MMD.RData", package="MMDiff2"))
runShinyMMDiff2(MMD)
}

}
